add user info helper
