+++ apps/nextjs-website/src/helpers/user.helper.ts
@@ -1,0 +1,38 @@
+/* eslint-disable functional/no-expression-statements */
+import { DevPortalUser } from '@/lib/types/auth';
+import { Auth, Hub } from 'aws-amplify';
+import { useCallback, useState, useEffect } from 'react';
+
+export const useUser = () => {
+  const [user, setUser] = useState<DevPortalUser | null>(null);
+
+  const checkUser = useCallback(() => {
+    Auth.currentAuthenticatedUser()
+      .then((user) => setUser(user))
+      .catch(() => setUser(null));
+  }, []);
+
+  useEffect(() => {
+    checkUser();
+  }, []);
+
+  useEffect(() => {
+    const cancel = Hub.listen('auth', (event) => {
+      switch (event.payload.event) {
+        case 'signIn':
+        case 'autoSignIn': {
+          const { data: user } = event.payload;
+          setUser(user);
+          break;
+        }
+        case 'signOut':
+          setUser(null);
+          break;
+      }
+    });
+
+    return () => cancel();
+  }, []);
+
+  return user;
+};
